[rails6][gemspec] Update dependencies
- Updates `graphql-batch` to the latest version
- Removes version restriction for `graphql-preload`

The latter was done due to that fact that `graphql-preload` currently
doesn't support Rails 6, and doesn't seem to be updated currently.

A fork of it, Envek/graphql-preload, has added support for Rails 6, so
for now that is what we will use.  However, to allow that to be
configured in the `Gemfile` in `ManageIQ/manageiq`, the version
restriction here needs to be loosened.